Tools and Agent bug fix
