+++ gwenflow/llms/response.py
@@ -18,7 +18,7 @@ class ModelResponse(BaseModel):
     thinking: Optional[str] = ""
     """The thinking of the response."""
 
-    # output: list[str]
+    output: list[str] = Field(default_factory=list)
     # """A list of outputs (messages, tool calls, etc) generated by the model"""
 
     finish_reason: Optional[str] = None
