+++ pyproject.toml
@@ -29,6 +29,7 @@ dependencies = [
     "pymupdf>=1.25.2",
     "asyncio>=3.4.3",
     "tenacity>=9.0.0",
+    "mcp>=1.6.0",
 ]
 
 classifiers = [
